+++ integration-tests/src/test/java/it/modules/TestInstallFailure.java
@@ -14,8 +14,8 @@ import com.atlassian.plugin.connect.modules.beans.builder.ConnectPageModuleBeanB
 import com.atlassian.plugin.connect.modules.beans.nested.I18nProperty;
 import com.atlassian.plugin.connect.modules.beans.nested.ScopeName;
 import com.atlassian.plugin.connect.test.AddonTestUtils;
+import com.atlassian.plugin.connect.test.pageobjects.ConnectAddOnEmbeddedTestPage;
 import com.atlassian.plugin.connect.test.pageobjects.LinkedRemoteContent;
-import com.atlassian.plugin.connect.test.pageobjects.RemotePluginEmbeddedTestPage;
 import com.atlassian.plugin.connect.test.server.ConnectRunner;
 import com.atlassian.upm.pageobjects.PluginManager;
 import com.google.common.collect.ImmutableMap;
@@ -23,7 +23,6 @@ import it.ConnectWebDriverTestBase;
 import it.servlet.ConnectAppServlets;
 import it.servlet.InstallHandlerServlet;
 import it.util.TestUser;
-
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -40,10 +39,9 @@ import java.security.NoSuchAlgorithmException;
 import static com.atlassian.plugin.connect.modules.beans.ConnectPageModuleBean.newPageBean;
 import static com.atlassian.plugin.connect.modules.util.ModuleKeyUtils.addonAndModuleKey;
 import static com.atlassian.plugin.connect.test.pageobjects.RemoteWebItem.ItemMatchingMode.LINK_TEXT;
-import static org.hamcrest.Matchers.equalTo;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 /**
@@ -76,12 +74,10 @@ public class TestInstallFailure extends ConnectWebDriverTestBase
                 .withWeight(1234);
         
         remotePlugin = new ConnectRunner(product.getProductInstance().getBaseUrl(), AddonTestUtils.randomAddOnKey())
-        .addInstallLifecycle()
         .addUninstallLifecycle()
         .addModule("configurePage", pageBeanBuilder.build())
-        .addJWT()
+        .addJWT(installUninstallHandler)
         .addRoute(route, ConnectAppServlets.helloWorldServlet())
-        .addRoute(ConnectRunner.INSTALLED_PATH, installUninstallHandler)
         .addRoute(ConnectRunner.UNINSTALLED_PATH, installUninstallHandler)
         .addScope(ScopeName.ADMIN)
         .disableInstallationStatusCheck();
@@ -166,9 +162,7 @@ public class TestInstallFailure extends ConnectWebDriverTestBase
                 "Configure",
                 Option.<String>none(), awesomePageModuleKey);
 
-        RemotePluginEmbeddedTestPage addonContentPage = addonPage.click();
-
-        assertThat(addonContentPage.isLoaded(), equalTo(true));
+        ConnectAddOnEmbeddedTestPage addonContentPage = addonPage.click();
 
         ConnectAsserts.verifyContainsStandardAddOnQueryParamters(addonContentPage.getIframeQueryParams(),
                 product.getProductInstance().getContextPath());