+++ integration-tests/src/test/java/it/modules/confluence/TestAdminPage.java
@@ -2,10 +2,9 @@ package it.modules.confluence;
 
 import com.atlassian.confluence.pageobjects.page.admin.ConfluenceAdminHomePage;
 import com.atlassian.plugin.connect.modules.beans.nested.I18nProperty;
-import com.atlassian.plugin.connect.modules.util.ModuleKeyUtils;
 import com.atlassian.plugin.connect.test.AddonTestUtils;
+import com.atlassian.plugin.connect.test.pageobjects.ConnectAddOnEmbeddedTestPage;
 import com.atlassian.plugin.connect.test.pageobjects.InsufficientPermissionsPage;
-import com.atlassian.plugin.connect.test.pageobjects.RemotePluginTestPage;
 import com.atlassian.plugin.connect.test.pageobjects.confluence.ConfluenceAdminPage;
 import com.atlassian.plugin.connect.test.pageobjects.confluence.ConnectConfluenceAdminHomePage;
 import com.atlassian.plugin.connect.test.server.ConnectRunner;
@@ -74,7 +73,7 @@ public class TestAdminPage extends ConfluenceWebDriverTestBase
     {
         loginAndVisit(TestUser.ADMIN, ConfluenceAdminHomePage.class);
 
-        ConfluenceAdminPage adminPage = product.getPageBinder().bind(ConfluenceAdminPage.class, ModuleKeyUtils.addonAndModuleKey(PLUGIN_KEY, PAGE_KEY));
+        ConfluenceAdminPage adminPage = product.getPageBinder().bind(ConfluenceAdminPage.class, PLUGIN_KEY, PAGE_KEY);
 
         assertThat(adminPage.isRemotePluginLinkPresent(), is(true));
 
@@ -83,7 +82,7 @@ public class TestAdminPage extends ConfluenceWebDriverTestBase
 
         // TODO Admin page web-item location has incorrect text ("OSGi")
 
-        RemotePluginTestPage addonContentsPage = adminPage.clickRemotePluginLink();
+        ConnectAddOnEmbeddedTestPage addonContentsPage = adminPage.clickAddOnLink();
         assertEquals("Hello world", addonContentsPage.getValueBySelector("#hello-world-message"));
     }