+++ modules/src/main/java/com/atlassian/plugin/connect/modules/beans/ModuleList.java
@@ -287,6 +287,14 @@ public class ModuleList extends BaseModuleBean
     @ConnectModule(value = "com.atlassian.plugin.connect.plugin.capabilities.provider.StaticContentMacroModuleProvider", products = {ProductFilter.CONFLUENCE})
     private List<StaticContentMacroModuleBean> staticContentMacros;
 
+    /**
+     * Blueprints allow your connect add on provide content creation templates.
+     *
+     * @schemaTitle Blueprint
+     */
+    @ConnectModule(value = "com.atlassian.plugin.connect.plugin.capabilities.provider.DefaultBlueprintModuleProvider", products = {ProductFilter.CONFLUENCE})
+    private List<BlueprintModuleBean> blueprints;
+
     public ModuleList()
     {
         this.adminPages = newArrayList();
@@ -305,6 +313,7 @@ public class ModuleList extends BaseModuleBean
         this.profilePages = newArrayList();
         this.spaceToolsTabs = newArrayList();
         this.staticContentMacros = newArrayList();
+        this.blueprints = newArrayList();
         this.webhooks = newArrayList();
         this.webItems = newArrayList();
         this.webPanels = newArrayList();
@@ -390,6 +399,10 @@ public class ModuleList extends BaseModuleBean
         if (null == staticContentMacros)
         {
             this.staticContentMacros = newArrayList();
+        }
+        if (null == blueprints)
+        {
+            this.blueprints = newArrayList();
         }
         if (null == jiraReports)
         {
@@ -501,6 +514,10 @@ public class ModuleList extends BaseModuleBean
         return staticContentMacros;
     }
 
+    public List<BlueprintModuleBean> getBlueprints() {
+        return blueprints;
+    }
+
     // don't call super because BaseCapabilityBean has no data
     @Override
     public boolean equals(Object otherObj)
@@ -534,10 +551,11 @@ public class ModuleList extends BaseModuleBean
                 .append(profilePages, other.profilePages)
                 .append(spaceToolsTabs, other.spaceToolsTabs)
                 .append(staticContentMacros, other.staticContentMacros)
+                .append(blueprints, other.blueprints)
                 .append(webhooks, other.webhooks)
                 .append(webItems, other.webItems)
-                .append(webPanels, webPanels)
-                .append(webSections, webSections)
+                .append(webPanels, other.webPanels)
+                .append(webSections, other.webSections)
                 .append(jiraReports, other.jiraReports)
                 .build();
     }
@@ -563,6 +581,7 @@ public class ModuleList extends BaseModuleBean
                 .append(profilePages)
                 .append(spaceToolsTabs)
                 .append(staticContentMacros)
+                .append(blueprints)
                 .append(webhooks)
                 .append(webItems)
                 .append(webPanels)