+++ plugin/src/main/resources/js/iframe/host/content-resolver.js
@@ -17,13 +17,13 @@ _AP.define('host/content-resolver', ['_dollar', '_ui-params'], function($, UiPar
             return $.ajax(getContentUrl(params.addonKey, params.moduleKey), {
                 dataType: "html",
                 data: {
-                    "ui-params": UiParams.encode(params),
+                    "ui-params": UiParams.encode(params.uiParams),
                     "plugin-key": params.addonKey,
                     "product-context": JSON.stringify(params.productContext),
                     "key": params.moduleKey,
-                    "width": "100%",
-                    "height": "100%",
-                    "raw": "true"
+                    "width": params.width || "100%",
+                    "height": params.height || "100%",
+                    "classifier": params.classifier || "raw"
                 }
             });
         }